Locking a 1-to-1  delivery session between ordering user and delivering user working

+++ api/services/OrderService.js
@@ -1,44 +1,54 @@
 module.exports = {
     deleteOrder: function (options, done) {
-        
-        Order.findOne({ id: options.req.param('id') }).exec(function (err, order) {
-            if (err) return options.res.negotiate(err);
-            if (!order) return options.res.notFound();
+        const thisUserId = options.req.session.userId;
+        const res = options.res;
 
-            // If this order was made by the user, delete it
-            if(order.userId == options.req.session.userId) {
-                sails.log.debug(`delete own order {OrderId:${order.id}, UserId:${options.req.session.userId}}`);                                              
-                Order.destroy({
-                    id: options.req.param('id'),
-                    userId: options.req.session.userId
-                }).exec(function (err, orderItems) {
-                    if (err) return options.res.negotiate(err);
-                    sails.log.debug(`delete own order successful {OrderId:${order.id}, UserId:${options.req.session.userId}}`);                                              
-                
-                    return options.res.json(orderItems);
-                });
-            } else if(order.deliverUserId == req.session.userId) {
-                // If this was made by the delivering user just clear them from delivery 
-                
-                sails.log.debug(`delete delivery user from order {OrderId:${order.id}, UserId:${options.req.session.userId}}`);                                              
-                
-                Order.update({
-                    deliverUserId: options.req.session.userId
-                }, {
-                    deliverUserId: null
-                }).exec(function(err, updatedOrder) {
-                    if (err) return options.res.negotiate(err);
-                    sails.log.debug(`delete delivery user successful {OrderId:${order.id}, UserId:${options.req.session.userId}}`);                                              
-                    
-                    return res.json(updatedOrder);
-                });
+        sails.log.debug(`delete order associated with {UserId:${thisUserId}}`);
+
+        // No need to check for ID, you will only have at most ONE order 
+        // associated with yourself in the orders table, delete this.
+
+        //Search database for existing order associated with this user
+        Order.findOne({
+            or: [
+                { userId: thisUserId },
+                { deliverUserId: thisUserId }
+            ]
+        }).exec(function (err, order) {
+            if (err) return res.negotiate(err);
+
+            // There is an order associated with this user
+            if (order) {
+                sails.log.debug(`user has an order to delete {OrderId:${order.id}, UserId:${thisUserId}}`);
+
+                if (order.userId == thisUserId) {
+                    // Clear any orders they have started
+                    Order.destroy({ userId: thisUserId }).exec(function (err, order) {
+                        if (err) return res.negotiate(err);
+                        if (order) {
+                            sails.log.debug(`delete own order successful {OrderId:${order.id}, UserId:${thisUserId}}`);
+                            return res.json(order);
+                        }
+                    });
+                } else {
+                    // Clear any orders they were delivering but were never completed
+                    Order.update({ deliverUserId: thisUserId }, { deliverUserId: null }).exec(function (err, order) {
+                        if (err) return res.negotiate(err);
+                        if (order) {
+                            sails.log.debug(`delete delivery user successful {OrderId:${order.id}, UserId:${thisUserId}}`);
+                            return res.json(order);
+                        }
+                    });
+                }
             } else {
-                // This delete request was made by neither the order user or delivery user
-                sails.log.debug(`user not allowed to delete order {OrderId:${order.id}, UserId:${options.req.session.userId}}`);                                              
-                return options.res.forbidden('You are not permitted to perform this action.');      
+                //No orders in here are associated with this user
+                sails.log.debug(`user has no order or delivery to delete {UserId:${thisUserId}}`);
+                return res.notModified();
             }
 
-                  
+
+
         });
+
     }
 }
